Add basic life-cycle handling:
This has a number of life-cycle phases that tasks are attached to.
It is intended for the user to interact only with the phases, each
has it's dependencies set so they are run in the correct order. Any
task attached to a phase will be run in parallel with the other tasks
in attached to the phase.